Move component into src folder
